feat: comprehensive strict typing implementation and test infrastructure overhaul
This commit implements comprehensive strict typing across the CogniVault platform
and resolves 35+ test failures through systematic cluster analysis and fixes.

Key Features:
- Complete Pydantic v2 migration with strict type validation
- Event system refactoring with proper WorkflowCompletedEvent validation
- Enhanced test infrastructure with factory patterns and proper isolation
- Performance optimizations for structured LLM calls
- API service layer improvements with proper error handling

Technical Improvements:
- Fixed circular import issues in BaseAgent event emission
- Implemented lazy loading for event functions to prevent import conflicts
- Enhanced test isolation with proper fixture management
- Added comprehensive factory classes for test data generation
- Improved MyPy compliance with proper generator return type annotations

Core Systems Enhanced:
- Agent orchestration with better error handling and retry logic
- Event-driven architecture with proper correlation tracking
- Database integration with Pydantic AI performance optimizations
- CLI and API boundary improvements with validation
- WebSocket event streaming with data integrity

Test Infrastructure:
- 139 files changed with 11,096 additions and 2,747 deletions
- Added 8+ new factory modules for comprehensive test coverage
- Resolved all MyPy strict typing errors (378 source files)
- Complete test suite passes reliably in all execution contexts
- Enhanced diagnostic tools and formatters with proper typing

This represents a major milestone in platform reliability and type safety,
establishing a solid foundation for production deployment and continued development.

+++ src/cognivault/events/sinks.py
@@ -106,7 +106,7 @@ class FileEventSink(EventSink):
     def __init__(
         self,
         file_path: str,
-        max_file_size_mb: int = 100,
+        max_file_size_mb: float = 100,
         rotate_files: bool = True,
         filters: Optional[EventFilters] = None,
     ) -> None:
@@ -264,7 +264,7 @@ class InMemoryEventSink(EventSink):
 # Factory function for creating file sinks
 def create_file_sink(
     file_path: str,
-    max_file_size_mb: int = 100,
+    max_file_size_mb: float = 100,
     rotate_files: bool = True,
     event_types: Optional[List[str]] = None,
     agent_names: Optional[List[str]] = None,
